ds 09: receipts rebuilt on the list-row recipe + semantic tokens
TransactionDetailsReceipt (1009 LoC) split into an orchestrator plus
ReceiptDetailsCard / ReceiptActions / ReceiptTokenRows / ReceiptReferralNudge
views, a ReceiptRow list-row primitive and two hooks (useTokenDisplay via
tanstack query, useReceiptActions for every charges/requests/claim-link call
so views import no api modules). The details card owns layout: px-4 +
divide-y dashed dividers, so rows and provider sub-rows carry no last-row
border logic — shouldHideBorder/shouldHideGroupBorder removed from the view
model. Provider rows (card payment, Manteca, Bridge deposit) and the perk
receipt moved off PaymentInfoRow onto ReceiptRow; legacy palette classes
replaced with semantic tokens. i18n keys unchanged.
