+++ src/components/TransactionDetails/ReceiptActions.tsx
@@ -1,0 +1,261 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import { useRouter } from 'next/navigation'
+import { useTranslations } from 'next-intl'
+import { Button } from '@/components/0_Bruddle/Button'
+import CancelSendLinkModal from '@/components/Global/CancelSendLinkModal'
+import { Icon } from '@/components/Global/Icons/Icon'
+import ShareButton from '@/components/Global/ShareButton'
+import { PasskeyDocsLink } from '@/components/Setup/Views/SignTestTransaction'
+import { CancelDepositActions } from './provider-actions/CancelDepositActions'
+import { ReceiptReferralNudge } from './ReceiptReferralNudge'
+import { ReceiptSupportLink } from './ReceiptSupportLink'
+import { type ReceiptViewModel } from './useReceiptViewModel'
+import { useReceiptActions } from './useReceiptActions'
+import { type TransactionDetails } from './transactionTransformer'
+import { hasReferralNudge, isRequestEntry, isSendLinkEntry, isSplittable } from './transaction-predicates'
+import { buildSplitBillRequestUrl } from './splitBill.utils'
+import { EHistoryUserRole } from '@/hooks/useTransactionHistory'
+import { useActivationStatus } from '@/hooks/useActivationStatus'
+import { useUserStore } from '@/redux/hooks'
+import { generateInviteCodeLink } from '@/utils/general.utils'
+import { getReceiptUrl, isTestTransaction } from '@/utils/history.utils'
+
+type CancelLinkState = 'idle' | 'cancelling' | 'cancelled'
+
+const CANCEL_LINK_KEYS = {
+    idle: 'actions.cancelLink',
+    cancelling: 'actions.cancelling',
+    cancelled: 'actions.cancelled',
+} as const satisfies Record<CancelLinkState, string>
+
+/**
+ * The receipt's CTA stack (DS 09): share/cancel for pending links, pay/reject
+ * for requests, split bill, share receipt, deposit cancels, referral nudge and
+ * the support footer. All api side effects route through useReceiptActions —
+ * this view only holds ephemeral UI state.
+ */
+export function ReceiptActions({
+    transaction,
+    vm,
+    isPublic,
+    amountDisplay,
+    shouldShowQrShare,
+    isLoading,
+    setIsLoading,
+    onClose,
+    setIsModalOpen,
+}: {
+    transaction: TransactionDetails
+    vm: ReceiptViewModel
+    isPublic: boolean
+    amountDisplay: string
+    shouldShowQrShare: boolean
+    isLoading?: boolean
+    setIsLoading?: (isLoading: boolean) => void
+    onClose?: () => void
+    setIsModalOpen?: (isModalOpen: boolean) => void
+}) {
+    const t = useTranslations('transaction')
+    const router = useRouter()
+    const { user } = useUserStore()
+    const { isActivated } = useActivationStatus()
+    const { closeRequest, rejectRequest, cancelSendLink } = useReceiptActions(transaction)
+    const { isPendingBankRequest, isPendingRequestee, isPendingRequester, isPendingSentLink, isQRPayment } = vm
+
+    const [showCancelLinkModal, setShowCancelLinkModal] = useState(false)
+    const [cancelLinkState, setCancelLinkState] = useState<CancelLinkState>('idle')
+
+    // Sync modal state to parent if callback is provided — the drawer keeps
+    // itself open while the confirm modal is up.
+    useEffect(() => {
+        setIsModalOpen?.(showCancelLinkModal)
+    }, [showCancelLinkModal, setIsModalOpen])
+
+    // `shouldShowShareReceipt` alone is TRUE for card spends (the txHash
+    // short-circuit in useReceiptViewModel); `getReceiptUrl` returning undefined
+    // is the real suppressor, so the CTA arithmetic must read the composite.
+    const receiptUrl = getReceiptUrl(transaction)
+    const showShareReceipt = vm.shouldShowShareReceipt && !!receiptUrl
+    const showSplitCta = !isPublic && isSplittable(transaction)
+
+    // `!isPublic`: on a public receipt the *viewer's* username would credit a
+    // bystander for someone else's payment.
+    const inviteUsername = user?.user.username
+    const showReferralNudge =
+        !isPublic &&
+        isActivated &&
+        transaction.status === 'completed' &&
+        hasReferralNudge(transaction) &&
+        !!inviteUsername
+    const inviteLink = inviteUsername ? generateInviteCodeLink(inviteUsername).inviteLink : ''
+
+    const referralCtaVariant = showSplitCta && showShareReceipt ? 'text_link' : 'button'
+
+    const handleCloseRequest = async () => {
+        if (!setIsLoading || !onClose) return
+        setIsLoading(true)
+        const ok = await closeRequest()
+        setIsLoading(false)
+        if (ok) onClose()
+    }
+
+    const handleRejectRequest = async () => {
+        if (!setIsLoading || !onClose) return
+        setIsLoading(true)
+        const ok = await rejectRequest()
+        setIsLoading(false)
+        if (ok) onClose()
+    }
+
+    const handleCancelSendLink = async () => {
+        if (!setIsLoading || !onClose) return
+        setIsLoading(true)
+        setCancelLinkState('cancelling')
+        const ok = await cancelSendLink()
+        setIsLoading(false)
+        if (!ok) {
+            setCancelLinkState('idle')
+            return
+        }
+        setShowCancelLinkModal(false)
+        setCancelLinkState('cancelled')
+        // Brief delay for toast visibility before the drawer closes.
+        await new Promise((resolve) => setTimeout(resolve, 1500))
+        onClose()
+    }
+
+    return (
+        <>
+            {/* share and cancel buttons section (only if qr is shown) */}
+            {shouldShowQrShare && transaction.extraDataForDrawer?.link && (
+                <div className="flex flex-col gap-2 pr-1">
+                    <ShareButton url={transaction.extraDataForDrawer.link} title={t('actions.shareLinkTitle')}>
+                        {t('actions.shareLink')}
+                    </ShareButton>
+                    {/* show cancel button only if the current user sent the link/request */}
+                    {(isSendLinkEntry(transaction) || isRequestEntry(transaction)) &&
+                        transaction.extraDataForDrawer.originalUserRole === EHistoryUserRole.SENDER &&
+                        setIsLoading &&
+                        onClose && (
+                            <Button
+                                disabled={isLoading || cancelLinkState === 'cancelled'}
+                                onClick={() => setShowCancelLinkModal(true)}
+                                loading={isLoading}
+                                variant="primary-soft"
+                                className="flex w-full items-center gap-1"
+                                shadowSize="4"
+                            >
+                                <div className="flex items-center">{!isLoading && <Icon name="ban" size={18} />}</div>
+                                <span>{t(CANCEL_LINK_KEYS[cancelLinkState])}</span>
+                            </Button>
+                        )}
+                </div>
+            )}
+
+            {isPendingSentLink && !shouldShowQrShare && (
+                <div className="flex items-center justify-center gap-1.5 text-center text-label-m text-foreground-secondary">
+                    <Icon name="info" size={20} />
+                    {t('pendingLinkDeviceNote')}
+                </div>
+            )}
+
+            {isPendingRequester && setIsLoading && onClose && (
+                <div className="pr-1">
+                    <Button
+                        icon="ban"
+                        iconSize={18}
+                        loading={isLoading}
+                        disabled={isLoading}
+                        onClick={handleCloseRequest}
+                        variant="primary-soft"
+                        shadowSize="4"
+                        className="flex w-full items-center gap-1"
+                    >
+                        {transaction.totalAmountCollected > 0 ? t('actions.closeRequest') : t('actions.cancelRequest')}
+                    </Button>
+                </div>
+            )}
+
+            {isPendingRequestee && setIsLoading && onClose && (
+                <div className="flex flex-col gap-2 pr-1">
+                    <Button
+                        onClick={() => {
+                            window.location.href = transaction.extraDataForDrawer?.link ?? ''
+                        }}
+                        shadowSize="4"
+                        className="flex w-full items-center gap-1"
+                    >
+                        <Icon name="currency" size={18} />
+                        {t('actions.pay')}
+                    </Button>
+                    <Button
+                        icon="ban"
+                        iconSize={18}
+                        disabled={isLoading}
+                        onClick={handleRejectRequest}
+                        variant="primary-soft"
+                        shadowSize="4"
+                        className="flex w-full items-center gap-1"
+                    >
+                        {t('actions.rejectRequest')}
+                    </Button>
+                </div>
+            )}
+
+            {showSplitCta && (
+                <Button
+                    onClick={() => router.push(buildSplitBillRequestUrl(transaction.amount, transaction.userName))}
+                    icon="split"
+                    shadowSize="4"
+                >
+                    {t('actions.splitBill')}
+                </Button>
+            )}
+
+            {showShareReceipt && (
+                <div className="pr-1">
+                    <ShareButton variant={isQRPayment ? 'primary-soft' : 'purple'} url={receiptUrl!}>
+                        {t('actions.shareReceipt')}
+                    </ShareButton>
+                </div>
+            )}
+
+            <CancelDepositActions
+                transaction={transaction}
+                isPendingBankRequest={isPendingBankRequest}
+                isLoading={isLoading}
+                setIsLoading={setIsLoading}
+                onClose={onClose}
+            />
+
+            {showReferralNudge && (
+                <ReceiptReferralNudge
+                    transactionId={transaction.id}
+                    inviteLink={inviteLink}
+                    variant={referralCtaVariant}
+                    label={t('actions.inviteFriends')}
+                />
+            )}
+
+            {/* support link section or passkey docs for test transactions */}
+            {isTestTransaction(transaction.userName) ? (
+                <PasskeyDocsLink className="border-t-0 pt-0" />
+            ) : (
+                <ReceiptSupportLink />
+            )}
+
+            {/* Cancel Link Modal */}
+            {setIsLoading && onClose && (
+                <CancelSendLinkModal
+                    showCancelLinkModal={showCancelLinkModal}
+                    setshowCancelLinkModal={setShowCancelLinkModal}
+                    amount={amountDisplay}
+                    isLoading={isLoading}
+                    onClick={handleCancelSendLink}
+                />
+            )}
+        </>
+    )
+}
