+++ src/components/TransactionDetails/ReceiptTokenRows.tsx
@@ -1,0 +1,75 @@
+'use client'
+
+import { useTranslations } from 'next-intl'
+import DisplayIcon from '@/components/Global/DisplayIcon'
+import { ReceiptRow } from '@/components/TransactionDetails/ReceiptRow'
+import { type TransactionDetails } from '@/components/TransactionDetails/transactionTransformer'
+import { useTokenDisplay } from '@/components/TransactionDetails/useTokenDisplay'
+import { isStableCoin } from '@/utils/general.utils'
+
+/**
+ * Token amount + token-and-network rows for the details card. Owns the token
+ * icon lookup (wire data first, CoinGecko fallback) via useTokenDisplay and
+ * renders nothing until icon + symbol resolve — same gate as the legacy
+ * receipt.
+ */
+export function ReceiptTokenRows({
+    transaction,
+    isPeanutWalletToken,
+}: {
+    transaction: TransactionDetails
+    isPeanutWalletToken: boolean
+}) {
+    const t = useTranslations('transaction')
+    const { tokenData, isLoading } = useTokenDisplay(transaction)
+
+    if (!transaction.tokenDisplayDetails || !tokenData?.icon || !tokenData?.symbol) return null
+
+    return (
+        <>
+            {!isStableCoin(transaction.tokenSymbol ?? 'USDC') && (
+                <ReceiptRow label={t('rows.tokenAmount')} value={transaction.amount} />
+            )}
+            {!isPeanutWalletToken && (
+                <ReceiptRow
+                    label={t('rows.tokenAndNetwork')}
+                    value={
+                        isLoading ? (
+                            <div className="h-6 w-32 animate-pulse rounded-sm bg-background-disabled" />
+                        ) : (
+                            <div className="flex items-center gap-2">
+                                <div className="relative flex h-6 w-6 min-w-[24px] items-center justify-center">
+                                    {/* Main token icon */}
+                                    <DisplayIcon
+                                        iconUrl={tokenData.icon}
+                                        altText={tokenData.symbol || 'token'}
+                                        fallbackName={tokenData.symbol || 'T'}
+                                        sizeClass="h-6 w-6"
+                                    />
+                                    {/* Smaller chain icon, absolutely positioned */}
+                                    {transaction.tokenDisplayDetails.chainIconUrl && (
+                                        <div className="absolute -right-1 -bottom-1">
+                                            <DisplayIcon
+                                                iconUrl={transaction.tokenDisplayDetails.chainIconUrl}
+                                                altText={transaction.tokenDisplayDetails.chainName || 'chain'}
+                                                fallbackName={transaction.tokenDisplayDetails.chainName || 'C'}
+                                                sizeClass="h-3.5 w-3.5 text-[7px]"
+                                                className="rounded-round border-2 border-background-default"
+                                            />
+                                        </div>
+                                    )}
+                                </div>
+                                <span>
+                                    {t('rows.tokenOnChain', {
+                                        token: tokenData.symbol.toUpperCase(),
+                                        chain: transaction.tokenDisplayDetails.chainName ?? '',
+                                    })}
+                                </span>
+                            </div>
+                        )
+                    }
+                />
+            )}
+        </>
+    )
+}
